solrj.getbeans can't process Java Bigint, so use long instead

+++ src/main/java/org/dataone/cn/batch/logging/listener/SystemMetadataEntryListener.java
@@ -209,7 +209,7 @@ public class SystemMetadataEntryListener implements EntryListener<Identifier, Sy
             solrItem.setIsPublic(isPublicSubject);
             solrItem.setReadPermission(subjectsAllowedRead);
             solrItem.setFormatId(systemMetadata.getFormatId().toString());
-            solrItem.setSize(systemMetadata.getSize());
+            solrItem.setSize(systemMetadata.getSize().longValue());
             solrItem.setRightsHolder(systemMetadata.getRightsHolder().toString());
             
             // Set the geographic location attributes determined from the IP address
